// src/modules/attendance/enums/attendance.enum.ts
export enum ENUM_ATTENDANCE_STATUS {
    PRESENT = 'present',
    ABSENT = 'absent',
    HALF_DAY = 'half_day',
    ON_LEAVE = 'on_leave',
    HOLIDAY = 'holiday',
}

export enum ENUM_ATTENDANCE_SOURCE {
    WEB = 'web',
    MOBILE = 'mobile',
    MANUAL = 'manual',
    AUTO = 'auto',
    KIOSK = 'kiosk',
}

export enum ENUM_BREAK_TYPE {
    LUNCH = 'lunch',
    SHORT = 'short',
    OTHER = 'other',
}